updates
updates in review services

+++ src/main/java/com/newgen/ReviewServiceApis/Model/Review.java
@@ -15,7 +15,7 @@ public class Review {
     private Long productId;
     private int userId;
 
-    private int ratings;
+    private short ratings;
     private String title;
     private String description;
